+++ packages/gen-typescript-declarations/src/test/goldens/paper-behaviors-3/paper-button-behavior.d.ts
@@ -8,7 +8,11 @@
  *   paper-button-behavior.js
  */
 
+import {IronButtonStateImpl, IronButtonState} from '@polymer/iron-behaviors/iron-button-state.js';
+
 import {PaperRippleBehavior} from './paper-ripple-behavior.js';
+
+import {IronControlState} from '@polymer/iron-behaviors/iron-control-state.js';
 
 export {PaperButtonBehaviorImpl};
 
@@ -69,4 +73,4 @@ export {PaperButtonBehavior};
 interface PaperButtonBehavior extends IronButtonState, IronControlState, PaperRippleBehavior, PaperButtonBehaviorImpl {
 }
 
-const PaperButtonBehavior: object;
+declare const PaperButtonBehavior: object;