+++ packages/gen-typescript-declarations/src/test/goldens/paper-behaviors-3/paper-inky-focus-behavior.d.ts
@@ -8,7 +8,11 @@
  *   paper-inky-focus-behavior.js
  */
 
+import {IronButtonState} from '@polymer/iron-behaviors/iron-button-state.js';
+
 import {PaperRippleBehavior} from './paper-ripple-behavior.js';
+
+import {IronControlState} from '@polymer/iron-behaviors/iron-control-state.js';
 
 export {PaperInkyFocusBehaviorImpl};