+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/legacy/templatizer-behavior.d.ts
@@ -8,7 +8,7 @@
  *   lib/legacy/templatizer-behavior.js
  */
 
-import {TemplateInstanceBase as TemplateInstanceBase$0, templatize as templatize$0, modelForElement as modelForElement$0} from '../utils/templatize.js';
+import {TemplateInstanceBase, templatize as templatize$0, modelForElement as modelForElement$0} from '../utils/templatize.js';
 
 export {Templatizer};
 
@@ -113,7 +113,7 @@ interface Templatizer {
   modelForElement(el: HTMLElement|null): TemplateInstanceBase|null;
 }
 
-const Templatizer: object;
+declare const Templatizer: object;
 
 
 /**