+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/mixins/gesture-event-listeners.d.ts
@@ -30,6 +30,8 @@ interface GestureEventListenersConstructor {
   new(...args: any[]): GestureEventListeners;
 }
 
+export {GestureEventListenersConstructor};
+
 interface GestureEventListeners {
 
   /**