+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/utils/async.d.ts
@@ -111,3 +111,7 @@ declare namespace microTask {
 }
 
 export {microTask};
+
+import {AsyncInterface} from '../../interfaces';
+
+import {IdleDeadline} from '../../interfaces';