+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/utils/debounce.d.ts
@@ -76,3 +76,5 @@ declare class Debouncer {
    */
   isActive(): boolean;
 }
+
+import {AsyncInterface} from '../../interfaces';