+++ packages/gen-typescript-declarations/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 - Support for ES module imports and exports.
 - Warnings are now printed with file names, line numbers, and code snippets.
+- Add `autoImport` config option to automatically add ES module imports when
+  particular identifiers are referenced.
+- Automatically detect if a project uses NPM or Bower and configure module
+  resolution settings accordingly.
+- Automatically import/export synthetic mixin constructor interfaces.
 
 ## [1.3.0] - 2018-06-29
 - Generate typings for class constructors.